requirements.txt
asgiref==3.5.0
coverage==6.3.2
Deprecated==1.2.13
Django==4.0.2
django-environ==0.8.1
django-filter==21.1
django-redis==5.2.0
djangorestframework==3.13.1
Faker==13.0.0
packaging==21.3
psycopg2==2.9.3
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.4
six==1.16.0
sqlparse==0.4.2
wrapt==1.13.3
